# docs/swagger.yaml
basePath: /
definitions:
  handlers.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  handlers.SupabaseLoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  handlers.SupabaseRegistroRequest:
    properties:
      data:
        additionalProperties: true
        type: object
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  models.APIError:
    properties:
      details:
        type: string
      message:
        type: string
      type:
        $ref: '#/definitions/models.ErrorType'
    type: object
  models.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/models.APIError'
    type: object
  models.ErrorType:
    enum:
    - validation_error
    - authentication_error
    - authorization_error
    - not_found_error
    - conflict_error
    - internal_error
    - bad_request_error
    type: string
    x-enum-varnames:
    - ErrorTypeValidation
    - ErrorTypeAuthentication
    - ErrorTypeAuthorization
    - ErrorTypeNotFound
    - ErrorTypeConflict
    - ErrorTypeInternal
    - ErrorTypeBadRequest
  models.EstatisticasProdutos:
    properties:
      cavaletes_disponiveis:
        type: integer
      produtos_destaque:
        type: integer
      produtos_visiveis:
        type: integer
      total_produtos:
        type: integer
    type: object
...
  models.ProdutoAprovado:
    properties:
      cavalete_id:
        type: string
      created_at:
        type: string
      descricao:
        type: string
      destaque:
        type: boolean
      id:
        type: string
      nome_customizado:
        maxLength: 255
        minLength: 1
        type: string
      ordem_exibicao:
        type: integer
      preco_venda:
        type: number
      trader_id:
        type: string
      updated_at:
        type: string
      visivel:
        type: boolean
    required:
    - nome_customizado
    - preco_venda
    type: object
  models.ProdutoAprovarRequest:
    properties:
      cavalete_id:
        type: string
      descricao:
        type: string
      destaque:
        type: boolean
      nome_customizado:
        maxLength: 255
        minLength: 1
        type: string
      preco_venda:
        type: number
      visivel:
        type: boolean
    required:
    - cavalete_id
    - nome_customizado
    - preco_venda
    type: object
  models.ProdutoAtualizarRequest:
    properties:
      descricao:
        type: string
      destaque:
        type: boolean
      nome_customizado:
        maxLength: 255
        minLength: 1
        type: string
      ordem_exibicao:
        type: integer
      preco_venda:
        type: number
      visivel:
        type: boolean
    type: object
  models.SupabaseAuthResponse:
    properties:
      session:
        $ref: '#/definitions/models.SupabaseSession'
      user:
        $ref: '#/definitions/models.SupabaseUser'
    type: object
  models.SupabaseSession:
    properties:
      access_token:
        type: string
      expires_at:
        type: integer
      refresh_token:
        type: string
    type: object
  models.SupabaseUser:
    properties:
      email:
        type: string
      id:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: "API para importação de ofertas do Mobgran com PostgreSQL - HOT RELOAD
    FUNCIONANDO! \U0001F525"
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Mobgran Importer API
  version: "1.0"
paths:
...
  /produtos/{id}:
    delete:
      description: Remove um produto aprovado
      parameters:
      - description: ID do produto
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Remover produto
      tags:
      - produtos
    get:
      description: Busca um produto específico do trader
      parameters:
      - description: ID do produto
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProdutoAprovado'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Buscar produto
      tags:
      - produtos
    put:
      consumes:
      - application/json
      description: Atualiza um produto aprovado
      parameters:
      - description: ID do produto
        in: path
        name: id
        required: true
        type: string
      - description: Dados para atualização
        in: body
        name: produto
        required: true
        schema:
          $ref: '#/definitions/models.ProdutoAtualizarRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProdutoAprovado'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Atualizar produto
      tags:
      - produtos
...
  /produtos/aprovar:
    post:
      consumes:
      - application/json
      description: Aprova um cavalete como produto para venda
      parameters:
      - description: Dados do produto
        in: body
        name: produto
        required: true
        schema:
          $ref: '#/definitions/models.ProdutoAprovarRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ProdutoAprovado'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Aprovar produto
      tags:
      - produtos
  /produtos/cavaletes:
    get:
      description: Lista cavaletes disponíveis para aprovação pelo trader
      parameters:
      - default: 20
        description: Limite de resultados
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset para paginação
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Listar cavaletes disponíveis
      tags:
      - produtos
  /produtos/estatisticas:
    get:
      description: Obtém estatísticas dos produtos do trader
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.EstatisticasProdutos'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Estatísticas de produtos
      tags:
      - produtos
  /supabase/auth/admin/create:
    post:
      consumes:
      - application/json
      description: Cria um novo usuário admin pré-confirmado usando Supabase Auth
      parameters:
      - description: Dados do usuário admin
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.SupabaseRegistroRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SupabaseAuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Criar usuário admin no Supabase
      tags:
      - supabase-auth
  /supabase/auth/login:
    post:
      consumes:
      - application/json
      description: Autentica um usuário usando Supabase Auth
      parameters:
      - description: Credenciais do usuário
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/handlers.SupabaseLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SupabaseAuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Login no Supabase
      tags:
      - supabase-auth
  /supabase/auth/logout:
    post:
      consumes:
      - application/json
      description: Faz logout do usuário no Supabase
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - supabase-auth
  /supabase/auth/refresh:
    post:
      consumes:
      - application/json
      description: Renova o token de acesso usando o refresh token
      parameters:
      - description: Refresh token
        in: body
        name: refresh
        required: true
        schema:
          $ref: '#/definitions/handlers.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SupabaseSession'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Renovar token
      tags:
      - supabase-auth
  /supabase/auth/user:
    get:
      consumes:
      - application/json
      description: Obtém informações do usuário autenticado no Supabase
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SupabaseUser'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Obter usuário atual
      tags:
      - supabase-auth
  /vitrine/publica:
    get:
      description: Lista produtos na vitrine pública (não requer autenticação)
      parameters:
      - default: 20
        description: Limite de resultados
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset para paginação
        in: query
        name: offset
        type: integer
      - description: Filtrar por trader específico
        in: query
        name: trader_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Vitrine pública
      tags:
      - produtos
schemes:
- http
- https
swagger: "2.0"